+++ src/main/java/com/cucumber/market/controller/ProductController.java
@@ -6,7 +6,10 @@ import com.cucumber.market.dto.member.CurrentMemberInfo;
 import com.cucumber.market.dto.product.ContentRequest;
 import com.cucumber.market.dto.product.ProductUploadForm;
 import com.cucumber.market.dto.product.ProductUploadResponse;
+import com.cucumber.market.service.CommentService;
+import com.cucumber.market.service.HopeService;
 import com.cucumber.market.service.ProductService;
+import com.cucumber.market.service.ReplyService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -27,6 +30,12 @@ import java.util.List;
 public class ProductController {
 
     private final ProductService productService;
+
+    private final HopeService hopeService;
+
+    private final CommentService commentService;
+
+    private final ReplyService replyService;
 
     // TODO: 2021-12-07 Springfox OpenAPI 3.0 MultiPartFile with JSON 적용 실패 이슈
     // JSON 형식으로 객체와 MultipartFile 파일을 RequestBody에 담으려 했으나 스웨거 테스트 하는 과정에서 이슈발생(깃허브 이슈등록)
@@ -73,8 +82,8 @@ public class ProductController {
     @CheckSignIn
     public ResponseEntity<Void> registerHope(@PathVariable int productIdx,
                                   @CurrentMember CurrentMemberInfo currentMemberInfo) {
-        productService.checkDuplicateHope(productIdx, currentMemberInfo.getMember_id());
-        productService.registerHope(productIdx, currentMemberInfo.getMember_id());
+        hopeService.checkDuplicateHope(productIdx, currentMemberInfo.getMember_id());
+        hopeService.registerHope(productIdx, currentMemberInfo.getMember_id());
 
         return ResponseEntity.ok().build();
     }
@@ -84,8 +93,8 @@ public class ProductController {
     @CheckSignIn
     public ResponseEntity<Void> cancelHope(@PathVariable int productIdx,
                                         @CurrentMember CurrentMemberInfo currentMemberInfo) {
-        productService.checkAlreadyHope(productIdx, currentMemberInfo.getMember_id());
-        productService.cancelHope(productIdx, currentMemberInfo.getMember_id());
+        hopeService.checkAlreadyHope(productIdx, currentMemberInfo.getMember_id());
+        hopeService.cancelHope(productIdx, currentMemberInfo.getMember_id());
 
         return ResponseEntity.ok().build();
     }
@@ -96,7 +105,7 @@ public class ProductController {
     public ResponseEntity<Void> registerComment(@PathVariable int productIdx,
                                              @Valid @RequestBody ContentRequest contentRequest,
                                              @CurrentMember CurrentMemberInfo currentMemberInfo) {
-        productService.registerComment(productIdx, contentRequest.getContent(), currentMemberInfo.getMember_id());
+        commentService.registerComment(productIdx, contentRequest.getContent(), currentMemberInfo.getMember_id());
         return ResponseEntity.ok().build();
     }
 
@@ -106,9 +115,9 @@ public class ProductController {
     public ResponseEntity<Void> updateComment(@PathVariable int commentIdx,
                                            @Valid @RequestBody ContentRequest request,
                                            @CurrentMember CurrentMemberInfo currentMemberInfo) {
-        productService.checkNotDeleteComment(commentIdx);
-        productService.checkCommentWriter(commentIdx, currentMemberInfo.getMember_id());
-        productService.updateComment(commentIdx, request.getContent());
+        commentService.checkNotDeleteComment(commentIdx);
+        commentService.checkCommentWriter(commentIdx, currentMemberInfo.getMember_id());
+        commentService.updateComment(commentIdx, request.getContent());
         return ResponseEntity.ok().build();
     }
 
@@ -117,9 +126,9 @@ public class ProductController {
     @CheckSignIn
     public ResponseEntity<Void> deleteComment(@PathVariable int commentIdx,
                                            @CurrentMember CurrentMemberInfo currentMemberInfo) {
-        productService.checkNotDeleteComment(commentIdx);
-        productService.checkCommentWriter(commentIdx, currentMemberInfo.getMember_id());
-        productService.deleteComment(commentIdx, currentMemberInfo.getMember_id());
+        commentService.checkNotDeleteComment(commentIdx);
+        commentService.checkCommentWriter(commentIdx, currentMemberInfo.getMember_id());
+        commentService.deleteComment(commentIdx, currentMemberInfo.getMember_id());
         return ResponseEntity.ok().build();
     }
 
@@ -131,8 +140,8 @@ public class ProductController {
                                            @Valid @RequestBody ContentRequest request,
                                            @CurrentMember CurrentMemberInfo currentMemberInfo) {
         // productIdx의 작성자이거나, commentIdx의 작성자인가
-        productService.checkProductOrCommentWriter(productIdx, commentIdx, currentMemberInfo.getMember_id());
-        productService.registerReply(commentIdx, request.getContent(), currentMemberInfo.getMember_id());
+        replyService.checkProductOrCommentWriter(productIdx, commentIdx, currentMemberInfo.getMember_id());
+        replyService.registerReply(commentIdx, request.getContent(), currentMemberInfo.getMember_id());
         return ResponseEntity.ok().build();
     }
 
@@ -142,9 +151,9 @@ public class ProductController {
     public ResponseEntity<Void> updateReply(@PathVariable int replyIdx,
                                          @Valid @RequestBody ContentRequest request,
                                          @CurrentMember CurrentMemberInfo currentMemberInfo) {
-        productService.checkNotDeleteReply(replyIdx);
-        productService.checkReplyWriter(replyIdx, currentMemberInfo.getMember_id());
-        productService.updateReply(replyIdx, request.getContent());
+        replyService.checkNotDeleteReply(replyIdx);
+        replyService.checkReplyWriter(replyIdx, currentMemberInfo.getMember_id());
+        replyService.updateReply(replyIdx, request.getContent());
         return ResponseEntity.ok().build();
     }
 
@@ -153,9 +162,9 @@ public class ProductController {
     @CheckSignIn
     public ResponseEntity<Void> deleteReply(@PathVariable int replyIdx,
                                          @CurrentMember CurrentMemberInfo currentMemberInfo) {
-        productService.checkNotDeleteReply(replyIdx);
-        productService.checkReplyWriter(replyIdx, currentMemberInfo.getMember_id());
-        productService.deleteReply(replyIdx, currentMemberInfo.getMember_id());
+        replyService.checkNotDeleteReply(replyIdx);
+        replyService.checkReplyWriter(replyIdx, currentMemberInfo.getMember_id());
+        replyService.deleteReply(replyIdx, currentMemberInfo.getMember_id());
         return ResponseEntity.ok().build();
     }
 }
