+++ src/main/java/com/cucumber/market/service/CommentService.java
@@ -1,0 +1,15 @@
+package com.cucumber.market.service;
+
+public interface CommentService {
+
+    void registerComment(int productIdx, String content, String member_id);
+
+    void checkNotDeleteComment(int commentIdx);
+
+    void checkCommentWriter(int commentIdx, String member_id);
+
+    void updateComment(int commentIdx, String content);
+
+    void deleteComment(int commentIdx, String member_id);
+
+}
