+++ src/main/java/com/cucumber/market/service/impl/ProductServiceImpl.java
@@ -3,9 +3,10 @@ package com.cucumber.market.service.impl;
 import com.cucumber.market.dto.product.FileUploadForm;
 import com.cucumber.market.dto.product.ProductUploadForm;
 import com.cucumber.market.dto.product.ProductUploadResponse;
-import com.cucumber.market.exception.*;
 import com.cucumber.market.file.FileStore;
-import com.cucumber.market.mapper.*;
+import com.cucumber.market.mapper.FileMapper;
+import com.cucumber.market.mapper.ProductMapper;
+import com.cucumber.market.mapper.ReplyMapper;
 import com.cucumber.market.service.ProductService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Value;
@@ -24,14 +25,6 @@ public class ProductServiceImpl implements ProductService {
     private final FileMapper fileMapper;
 
     private final FileStore fileStore;
-
-    private final HopeMapper hopeMapper;
-
-    private final CommentMapper commentMapper;
-
-    private final ReplyMapper replyMapper;
-
-
 
     @Value("${cucumber.product.url}")
     private String productUrl;
@@ -90,168 +83,5 @@ public class ProductServiceImpl implements ProductService {
                 .build();
     }
 
-    /**
-     * 상품찜 메소드
-     * @param productIdx 상품번호
-     * @param member_id 로그인한 회원의 아이디
-     */
-    @Override
-    public void registerHope(int productIdx, String member_id) {
-        hopeMapper.registerHope(productIdx, member_id);
-    }
 
-    /**
-     * 상품찜 중복검사 메소드
-     * @param productIdx 상품번호
-     * @param member_id 로그인한 회원의 아이디
-     */
-    @Override
-    public void checkDuplicateHope(int productIdx, String member_id) {
-        if (hopeMapper.checkDuplicateHope(productIdx, member_id) == 1) {
-            throw new AlreadyHopeRegisteredException("이미 해당 상품을 찜하셨습니다.");
-        }
-    }
-
-    /**
-     * 상품찜 취소전 본인확인 메소드
-     * @param productIdx 상품번호
-     * @param member_id 로그인한 회원의 아이디
-     */
-    @Override
-    public void checkAlreadyHope(int productIdx, String member_id) {
-        if (hopeMapper.checkDuplicateHope(productIdx, member_id) == 0) {
-            throw new NotYetHopeRegisteredException("해당 상품을 찜한적이 없습니다.");
-        }
-    }
-
-    /**
-     * 상품찜 취소 메소드
-     * @param productIdx 상품번호
-     * @param member_id 로그인한 회원의 아이디
-     */
-    @Override
-    public void cancelHope(int productIdx, String member_id) {
-        hopeMapper.cancelHope(productIdx, member_id);
-    }
-
-    /**
-     * 댓글등록 메소드
-     * @param productIdx 상품번호
-     * @param content 댓글내용
-     * @param member_id 로그인한 회원의 아이디
-     */
-    @Override
-    public void registerComment(int productIdx, String content, String member_id) {
-        commentMapper.registerComment(productIdx, content, member_id);
-    }
-
-    /**
-     * 댓글삭제 여부검사 메소드
-     * @param commentIdx 댓글번호
-     */
-    @Override
-    public void checkNotDeleteComment(int commentIdx) {
-        if (commentMapper.checkNotDeleteComment(commentIdx) == 1) {
-            throw new AlreadyDeletedCommentException("해당 댓글은 이미 삭제되었습니다.");
-        }
-    }
-
-    /**
-     * 댓글 작성자 확인 메소드
-     * @param commentIdx 댓글번호
-     * @param member_id 로그인한 회원의 아이디
-     */
-    @Override
-    public void checkCommentWriter(int commentIdx, String member_id) {
-        if (commentMapper.checkCommentWriter(commentIdx, member_id) == 0) {
-            throw new NoWriterForCommentException("해당 댓글의 작성자가 아닙니다.");
-        }
-    }
-
-    /**
-     * 댓글수정 메소드
-     * @param commentIdx 댓글번호
-     * @param content 수정할 내용
-     */
-    @Override
-    public void updateComment(int commentIdx, String content) {
-        commentMapper.updateComment(commentIdx, content);
-    }
-
-    /**
-     * 댓글삭제 메소드
-     * @param commentIdx 댓글번호
-     * @param member_id 로그인한 회원의 아이디
-     */
-    @Override
-    public void deleteComment(int commentIdx, String member_id) {
-        commentMapper.deleteComment(commentIdx, member_id);
-    }
-
-    /**
-     * 대댓글 등록 메소드
-     * @param commentIdx 댓글번호
-     * @param content 등록할 내용
-     * @param member_id 로그인한 회원의 아이디
-     */
-    @Override
-    public void registerReply(int commentIdx, String content, String member_id) {
-        replyMapper.registerReply(commentIdx, content, member_id);
-    }
-
-    /**
-     * 상품글 작성자 or 댓글 작성자 검사 메소드
-     * @param productIdx 글번호
-     * @param commentIdx 댓글번호
-     * @param member_id 대댓글 작성을 시도하는 회원의 아이디
-     */
-    @Override
-    public void checkProductOrCommentWriter(int productIdx, int commentIdx, String member_id) {
-        if (productMapper.checkProductWriter(productIdx, member_id) == 0 || commentMapper.checkCommentWriter(commentIdx, member_id) == 0) {
-            throw new NoWriterForProductOrCommentException("해당 글이나 댓글의 작성자가 아닙니다.");
-        }
-    }
-
-    /**
-     * 삭제된 대댓글인지 확인하는 메소드
-     * @param replyIdx 대댓글 번호
-     */
-    @Override
-    public void checkNotDeleteReply(int replyIdx) {
-        if (replyMapper.checkNotDeleteReply(replyIdx) == 1) {
-            throw new AlreadyDeletedReplyException("해당 대댓글은 이미 삭제되었습니다.");
-        }
-    }
-
-    /**
-     * 대댓글 작성자인지 확인하는 메소드
-     * @param replyIdx 대댓글 번호
-     * @param member_id 로그인한 회원의 아이디
-     */
-    @Override
-    public void checkReplyWriter(int replyIdx, String member_id) {
-        if (replyMapper.checkReplyWriter(replyIdx, member_id) == 0) {
-            throw new NoWriterForReplyException("해당 대댓글의 작성자가 아닙니다.");
-        }
-    }
-
-    /**
-     * 대댓글 수정 메소드
-     * @param replyIdx 대댓글 번호
-     * @param content 수정할 내용
-     */
-    @Override
-    public void updateReply(int replyIdx, String content) {
-        replyMapper.updateReply(replyIdx, content);
-    }
-
-    /**
-     * 대댓글 삭제 메소드
-     * @param replyIdx 대댓글 번호
-     * @param member_id 로그인한 회원의 아이디
-     */
-    @Override
-    public void deleteReply(int replyIdx, String member_id) {
-        replyMapper.deleteReply(replyIdx, member_id);
-    }
 }
